tests/wpt/web-platform-tests/webmessaging/without-ports/025.html

<!doctype html>
<title>MessagePort constructor properties</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
test(function() {
   var ch = new MessageChannel();
   assert_true(ch.port1 instanceof MessagePort, "MessageChannel's port not an instance of MessagePort");
   assert_throws(new TypeError(), function () { var p = new MessagePort();}, "MessagePort is [[Callable]]");
});
fetch_tests_from_worker(new Worker("025-1.js"));
</script>
